requirements.txt
streamlit
python-pptx
pandas
openpyxl
